Rename text state methods
